Render text nodes via their parent element
This patch fixes a bug I experienced when working with contenteditable
elements.

It turns out that deleting the text of a contenteditable element
causes the text node to be removed from the DOM and replaced with
a <br> element instead.

domChanger apparently does not expect text nodes to ever be removed by
the browser, presumably because the contenteditable attribute either
didn't exist or wasn't considered at the time of writing.  So it
continues to store a reference to the text node in its tree after it
has been removed from the DOM.  It will continue to update this node
on subsequent renders, but the user will never see them as the node
is no longer attached.

This patch fixes that issue by not storing any references to text nodes
at all, but instead by altering their contents using their parents'
innerText property.

This should work fine as there is only ever one text node per element
and its value should always be the same as the value of innerText.